+++ projects/demo-application/src/app/app.module.ts
@@ -9,18 +9,22 @@ import {OutputComponent} from './components/output/output.component';
 import {MatCardModule} from "@angular/material/card";
 import {MatNativeDateModule} from "@angular/material/core";
 import {MatButtonModule} from "@angular/material/button";
+import {MatMenuModule} from "@angular/material/menu";
+import { QuerySearchComponent } from './components/query-search/query-search.component';
+import {MatIconModule} from "@angular/material/icon";
 
 @NgModule({
   declarations: [
     AppComponent,
-    OutputComponent
+    OutputComponent,
+    QuerySearchComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule,
     QuerySearchModule.forRoot({
-      loggingCallback:  (...args) => {
+      loggingCallback: (...args) => {
         console.log(...args);
       },
       debug: true,
@@ -34,7 +38,9 @@ import {MatButtonModule} from "@angular/material/button";
     MatToolbarModule,
     MatCardModule,
     MatNativeDateModule,
-    MatButtonModule
+    MatButtonModule,
+    MatMenuModule,
+    MatIconModule
   ],
   bootstrap: [AppComponent],
 })