+++ projects/query-search/src/lib/components/query-search-group/query-search-group.component.ts
@@ -1,11 +1,19 @@
-import {Component, EventEmitter, HostBinding, Input, Output, QueryList, ViewChildren} from '@angular/core';
-import {QueryGroup, SavedFilter} from "../../models";
+import {
+  Component,
+  EventEmitter,
+  HostBinding,
+  Input,
+  Output,
+  QueryList,
+  ViewChildren
+} from '@angular/core';
+import {QueryGroup, QueryItem, SavedFilter} from "../../models";
 import {QuerySearchService} from "../../query-search.service";
 import {inOutAnimations} from "../../animations";
 import {QuerySearchItemComponent} from "../query-search-item/query-search-item.component";
 import {CdkDragDrop, moveItemInArray} from "@angular/cdk/drag-drop";
-import {NameDialogComponent} from "../dialogs/name-dialog/name-dialog.component";
 import {isEven} from "../../helpers/general.helpers";
+import {MatMenu} from "@angular/material/menu";
 
 @Component({
   selector: 'query-search-group',
@@ -17,8 +25,11 @@ import {isEven} from "../../helpers/general.helpers";
 })
 export class QuerySearchGroupComponent {
 
-  @Output()
-  generateClicked = new EventEmitter<boolean>(true);
+  @Input() filterMenu?: MatMenu;
+
+  @Output() generateClicked = new EventEmitter<boolean>(true);
+
+  @Output() filterCleared = new EventEmitter();
 
   @Input()
   set group(newValue: QueryGroup) {
@@ -31,6 +42,8 @@ export class QuerySearchGroupComponent {
   @ViewChildren(QuerySearchItemComponent) items: QueryList<QuerySearchItemComponent>;
   @ViewChildren(QuerySearchGroupComponent) children: QueryList<QuerySearchGroupComponent>;
 
+  currentFilter?: SavedFilter;
+  currentFilterChanged = false;
   loadingFilters = new EventEmitter<boolean>(true);
 
   get group() {
@@ -65,28 +78,34 @@ export class QuerySearchGroupComponent {
   addItem() {
     this.group.addItem();
     this.querySearchService.log('Adding item to', this);
+    this.currentFilterChanged = true;
   }
 
   addGroup() {
     this.group.addChild();
     this.querySearchService.log('Adding group to', this);
+    this.currentFilterChanged = true;
   }
 
   removeItem(id: string) {
     this.group.removeItem(id);
     this.querySearchService.log('Removed item with ID', id, 'from', this);
+    this.currentFilterChanged = true;
   }
 
   removeChild(id: string) {
     this.group.removeChild(id);
     this.querySearchService.log('Removed child with ID', id, 'from', this);
+    this.currentFilterChanged = true;
   }
 
-  remove() {
+  remove(emit: boolean = true) {
     this.querySearchService.log('Removing group', this.group.id, this);
-    this.children.forEach(child => child.remove());
-    this.items.forEach(item => item.remove());
-    this.removed.emit(this.group.id);
+    this.children.forEach(child => child.remove(emit));
+    this.items.forEach(item => item.remove(emit));
+    if (emit) {
+      this.removed.emit(this.group.id);
+    }
   }
 
   clear() {
@@ -94,57 +113,32 @@ export class QuerySearchGroupComponent {
     this.children.forEach(child => child.remove());
     this.items.forEach(item => item.remove());
     this.group.addItem();
+    this.currentFilter = undefined;
+    this.currentFilterChanged = false;
+    this.filterCleared.emit(undefined);
   }
 
   drop(event: CdkDragDrop<string[]>) {
     moveItemInArray(this.group.items, event.previousIndex, event.currentIndex);
+    this.currentFilterChanged = true;
   }
 
-  loadFilter(filterName: string) {
+  loadFilter(filter: SavedFilter | null) {
     this.loadingFilters.emit(true);
-
-    const filter = this.getFilter(filterName);
+    this.currentFilterChanged = false;
 
     if (!!filter) {
-      this.children.forEach(child => child.remove());
-      this.items.forEach(item => item.remove());
+      this.children.forEach(child => child.remove(false));
+      this.items.forEach(item => item.remove(false));
       this.group.apply(filter.ruleGroup);
-
-      this.loadingFilters.emit(false);
-    } else {
-      this.loadingFilters.emit(false);
+      this.currentFilter = filter;
     }
+
+    this.loadingFilters.emit(false);
   }
 
-  renameFilter(filterName: string) {
-    const filter = this.getFilter(filterName);
-
-    if (!!filter) {
-      this.querySearchService.openFilterNameDialog(NameDialogComponent, filter, 'EDIT').subscribe();
-    }
-  }
-
-  deleteFilter(filterName: string) {
-    const filter = this.getFilter(filterName);
-    if (!!filter) {
-      this.querySearchService.deleteFilter(filter);
-    }
-  }
-
-  createFilter() {
-    const queryRuleGroup = this.group.filterValue;
-
-    if (Object.keys(queryRuleGroup).length === 0) {
-      this.querySearchService.log('Refusing to make filter from empty QueryRuleGroup', queryRuleGroup);
-      return;
-    }
-
-    const savedFilter: SavedFilter = {
-      name: '',
-      ruleGroup: queryRuleGroup
-    }
-
-    this.querySearchService.openFilterNameDialog(NameDialogComponent, savedFilter, 'CREATE').subscribe();
+  markChanged(item: QueryItem) {
+    this.currentFilterChanged = true;
   }
 
 
@@ -182,17 +176,5 @@ export class QuerySearchGroupComponent {
 
   get isTopLevel(): boolean {
     return this.group.depth === 0;
-  }
-
-  private getFilter(filterName: string) {
-    const filter = this.querySearchService.savedFilters.value.find(f => f.name === filterName);
-
-    if (!!filter) {
-      return filter
-    }
-
-    this.querySearchService.log('Could not find filter for name', filterName);
-
-    return null;
   }
 }