+++ projects/query-search/src/lib/models/name-dialog-data.model.ts
@@ -1,6 +1,0 @@
-import {SavedFilter} from "./saved-filter.model";
-
-export interface NameDialogData {
-  action: 'EDIT' | 'CREATE';
-  filter: SavedFilter;
-}