+++ projects/query-search/src/lib/models/value-notification.model.ts
@@ -1,6 +1,0 @@
-import {QueryField} from "./query-field.model";
-
-export interface ValueNotification {
-  partialValue: string;
-  field: QueryField;
-}